pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pydocstruct"
dynamic = ["version"]
description = "Document structuring library for RAG and vector databases"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["rag", "document", "vector", "embedding", "nlp"]
authors = [
  {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
]

dependencies = [
  "pypdf>=3.0.0",
  "python-docx>=1.0.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=4.9.0",
  "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "black>=23.0.0",
  "flake8>=6.0.0",
  "mypy>=1.5.0",
  "isort>=5.12.0",
]
all = [
  "openpyxl>=3.1.0",
  "pandas>=2.0.0",
  "tabulate>=0.9.0",
  "markdownify>=0.11.0",
  "tiktoken>=0.5.0",
  "pdf2image>=1.16.0",
  "pytesseract>=0.3.10",
]

[project.urls]
Homepage = "https://github.com/shintaro-amaike/pydocstruct"
Documentation = "https://docstruct.readthedocs.io"
Repository = "https://github.com/shintaro-amaike/pydocstruct"
"Bug Tracker" = "https://github.com/shintaro-amaike/pydocstruct/issues"

[tool.black]
line-length = 79
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=pydocstruct --cov-report=html --cov-report=term"